% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annuity.R
\name{annuity}
\alias{annuity}
\title{Compute investment growth with annuities}
\usage{
annuity(return_rate, initial_investment, annuity, years)
}
\arguments{
\item{return_rate}{A numeric representing the annual return rate.}

\item{initial_investment}{A numeric representing the initial investment amount.}

\item{annuity}{A numeric representing the amount added to the investment every year.}

\item{years}{A numeric representing the number of years of the investment period.}
}
\value{
A numeric vector representing the size of the investment each year.
}
\description{
This function takes an initial investment amount, an amount for annuities
added to the investment every year, an annual return rate, and years of investment,
and returns a numeric vector representing the new amount every year.
}
\examples{
# With an initial investment of $10,000 and yearly contributions of $5,000 ...
# ... what will the size of the investment be after 25 years of 8\% annual returns?
annuity(.08,10000,5000,25)

}
