% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annuity_n_returns.R
\name{annuity_n_returns}
\alias{annuity_n_returns}
\title{Compute investment growth with annuities for a variety of annual return rates}
\usage{
annuity_n_returns(
  initial_investment,
  annuity,
  years,
  return_rates = c(0.06, 0.07, 0.08, 0.09, 0.1)
)
}
\arguments{
\item{initial_investment}{A numeric representing the initial investment amount.}

\item{annuity}{A numeric representing the amount added to the investment every year.}

\item{years}{A numeric representing the number of years of the investment period.}

\item{return_rates}{A numeric vector representing the specified annual return rates.}
}
\value{
A list with: 1) a plotly plot object displaying the growth curves of the various investments, and 2) a tibble containing the data displayed in the plot.
}
\description{
This function takes an initial investment amount, an amount for annuities
added to the investment every year, a vector of annual return rates, and years of investment,
and returns an interactive plot displaying the performance of the various return rates
over the specified period, as well as a tibble with the data displayed in the plot.
}
\examples{
# How do 6\%, 7\%, 8\%, 9\%, and 10\% annual return rates compare after ...
# ... 25 years of an initial $20,000 investment, & $10,000 annual contributions?
list_data_and_plot <- annuity_n_returns(
 20000,
 10000,
 25,
 c(.06,.07,.08,.09,.1)
)

list_data_and_plot[[1]]

}
