% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_dist.R
\name{binomial_dist}
\alias{binomial_dist}
\title{Plot the distribution of a specified binomial random variable, with probability of specified event}
\usage{
binomial_dist(n_trials, prob, tail_cutoff, left_or_right)
}
\arguments{
\item{n_trials}{A non-zero integer representing the number of Bernoulli trials in the experiment.}

\item{prob}{A numeric value representing the probability of success for each trial. Range: 0 to 1.}

\item{tail_cutoff}{An integer representing the cutoff point for the number of successes. To the right or left of this cutoff (inclusive) is the desired range or the portion of the distribution that corresponds to the event of interest.}

\item{left_or_right}{A character indicating whether your event of interest corresponds to the possible range of successes to the right or to the left of your cutoff point. Possible values: c("left", "right").}
}
\value{
An interactive plotly plot with the specified binomial distribution, a portion of the distribution colored in
to represent the event of interest, and the probability of the event in the plot title.
}
\description{
This function takes parameters for a binomial experiment (n trials, probability of success)
and a tail cutoff point for the tail of the distribution that represents the event for which
you want to know the probability (an event consisting of a range of values for the number of
successes), and plots the distribution with the portion corresponding to the event filled in.
}
\examples{
# If I apply to 8 schools, each with an acceptance rate of 20\% ...
# ... what is the probability that I get into at least one of them?
binomial_dist(8,.20,1,"right")

# 55 people received a particular surgery, were followed to see if they regretted it.
# Zero regretted it. If you were expecting 5\% of them to regret it, then ...
# ... under that assumption what was the probability zero would end up regretting it?
binomial_dist(55,.05,0,"left")

}
