% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cali_tbl}
\alias{cali_tbl}
\title{California State Tax Brackets}
\format{
A data frame with 9 rows and 3 columns:
\describe{
\item{rate}{The marginal tax rate}
\item{joint_lb}{The tax bracket lower bound for married filing jointly that goes with the corresponding marginal tax rate}
\item{joint_ub}{The tax bracket upper bound for married filing jointly that goes with the corresponding marginal tax rate}
}
}
\usage{
cali_tbl
}
\description{
This dataset contains the state tax rate of California for each tax bracket.
}
\keyword{datasets}
