% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_normality.R
\name{check_normality}
\alias{check_normality}
\title{Asses normality using a QQ plot, a histogram with normal density curve superimposed, and a Shapiro-Wilk test for normality}
\usage{
check_normality(data, variable_name, just_qqplot = F, just_shapiro_pval = F)
}
\arguments{
\item{data}{A data frame with the variable of interest.}

\item{variable_name}{A character; the name of the variable of interest. The variable's sample size must be between 3 and 5,000, inclusive.}

\item{just_qqplot}{A logical; do you want just the qqplot? TRUE/FALSE. Default is FALSE.}

\item{just_shapiro_pval}{A logical; do you want just the Shapiro-Wilk p-value? TRUE/FALSE. Default is FALSE.}
}
\value{
A qqplot, a histogram of the variable's distribution with a normal curve superimposed, and the results of a Shapiro-Wilk test.
}
\description{
This function takes a data frame and a string of the variable name for
which you wish to assess normality, and outputs the QQ plot, the histogram with
superimposed normal curve, and the results of the Shapiro-Wilk test.
}
\examples{
# Assess the normality of mpg in mtcars
check_normality(ChickWeight,"weight")

}
