% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fed_tbl}
\alias{fed_tbl}
\title{Federal Tax Brackets}
\format{
A data frame with 7 rows and 7 columns:
\describe{
\item{rate}{The marginal tax rate}
\item{single_lb}{The tax bracket lower bound for filing single that goes with the corresponding marginal tax rate}
\item{single_ub}{The tax bracket upper bound for filing single that goes with the corresponding marginal tax rate}
\item{joint_lb}{The tax bracket lower bound for married filing jointly that goes with the corresponding marginal tax rate}
\item{joint_ub}{The tax bracket upper bound for married filing jointly that goes with the corresponding marginal tax rate}
\item{head_lb}{The tax bracket lower bound for filing head of household that goes with the corresponding marginal tax rate}
\item{head_ub}{The tax bracket upper bound for filing head of household that goes with the corresponding marginal tax rate}
}
}
\usage{
fed_tbl
}
\description{
This dataset contains the federal tax rate for each tax bracket.
}
\keyword{datasets}
