% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{okla_tbl}
\alias{okla_tbl}
\title{Oklahoma State Tax Brackets}
\format{
A data frame with 7 rows and 5 columns:
\describe{
\item{rate}{The marginal tax rate}
\item{single_lb}{The tax bracket lower bound for filing single that goes with the corresponding marginal tax rate}
\item{single_ub}{The tax bracket upper bound for filing single that goes with the corresponding marginal tax rate}
\item{joint_lb}{The tax bracket lower bound for married filing jointly that goes with the corresponding marginal tax rate}
\item{joint_ub}{The tax bracket upper bound for married filing jointly that goes with the corresponding marginal tax rate}
}
}
\usage{
okla_tbl
}
\description{
This dataset contains the state tax rate of Oklahoma for each tax bracket.
}
\keyword{datasets}
