% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_logistic_results.R
\name{plot_logistic_results}
\alias{plot_logistic_results}
\title{Dot-whisker plot of odds ratios and 95\% confidence intervals from a multivariable logistic regression model}
\usage{
plot_logistic_results(
  model,
  title = "Adjusted Odds Ratios with 95\% CI",
  subtitle = NULL,
  log_scale = TRUE
)
}
\arguments{
\item{model}{A fitted \code{glm} model.}

\item{title}{A character, by default set to 'Adjusted Odds Ratios with 95\% CI'.}

\item{subtitle}{A character. Default is NULL.}

\item{log_scale}{A logical. Would you like to see the odds ratios on a log scale? TRUE/FALSE. Default is TRUE.}
}
\value{
A dot-whisker plot of the estimated adjusted odds ratios with 95\% confidence intervals.
}
\description{
This function takes a glm object (the multivariable logistic regression model) and returns a ggplot of the
estimated adjusted odds ratios for each covariate, and corresponding 95\% confidence intervals.
}
\examples{
# Visualize the relative odds of not surviving the Titanic
# E.g., For males vs. females, after adjusting for class and age group
library(dplyr)

data("Titanic")

titanic_tbl <- Titanic \%>\%
tibble::as_tibble() \%>\%
tidyr::uncount(n) \%>\%
mutate(
  Survived = factor(Survived,levels = c("Yes","No")) \%>\% as.numeric(),
  Survived = Survived - 1,
  Survived = factor(Survived)
 )

titanic_fit <- glm(Survived ~ Sex + Age + Class,data = titanic_tbl,family = binomial)

plot_logistic_results(
titanic_fit,
subtitle = "Relative odds of not surviving the Titanic"
)

}
