% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_na_cols.R
\name{remove_na_cols}
\alias{remove_na_cols}
\title{Remove columns in a data frame that only have NA values (or that exceed a specified proportion of NA values)}
\usage{
remove_na_cols(data, threshold = 1)
}
\arguments{
\item{data}{A data frame or tibble. If there are NA-only columns, they will be removed, though it isn't necessary for data to have NA-only columns.}

\item{threshold}{A numeric between 0 and 1. Default is 1 (i.e., 100\%). Represents the threshold proportion of NA values if the user wants to remove columns that are not purely NA, but whose proportion of NA values exceeds the specified threshold.}
}
\value{
Your data frame, minus the columns that were removed
}
\description{
This function takes a data frame and automatically removes the columns
that only contain NA values. Also, the user can set a desired threshold for removing
columns with a proportion of NA values exceeding that threshold.
}
\examples{
# Remove columns that are NA-only
library(dplyr)

mtcars$drat <- NA_real_
mtcars_new <- mtcars \%>\% remove_na_cols()
mtcars_new

# Set a threshold proportion of NA values for column-removal
library(dplyr)

mtcars$drat <- NA_real_
mtcars \%>\%
  mutate(drat = case_when(mpg > 31 ~ 3.5)) \%>\%
  remove_na_cols(threshold = .93) # Change threshold to .94 and drat will remain

}
