% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_proportion_plot.R
\name{strat_proportion_plot}
\alias{strat_proportion_plot}
\title{Plot bivariate proportions with confidence intervals}
\usage{
strat_proportion_plot(
  data,
  row_var,
  col_var,
  x_label = "Row Variable",
  plot_title = "Stratified Counts and Percentages"
)
}
\arguments{
\item{data}{A data frame.}

\item{row_var}{The categorical row variable (not passed as a string).}

\item{col_var}{The categorical column variable (not passed as a string). The counts and percents of the row variable will be stratified by the column variable.}

\item{x_label}{A character. The X axis label.}

\item{plot_title}{A character. The plot title.}
}
\value{
A ggplot of the stratified proportions with CI errorbars
}
\description{
This function takes a data frame, a categorical row variable, and a categorical column variable, and
plots the proportions of the row variable categories stratified by (faceted by) the column variable categories.
Error bars represent the 95\% confidence intervals for the proportions.
}
\examples{
# Plot of counts and \% of cylinder categories stratified by transmission category in mtcars
strat_proportion_plot(mtcars,cyl,am,"Cylinder","Number of cylinders (count, \%) by transmission type")

}
