% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxes.R
\name{taxes}
\alias{taxes}
\title{Compute your net income from your gross income, after factoring in federal and state taxes}
\usage{
taxes(gross_income, state_name = "OK")
}
\arguments{
\item{gross_income}{A numeric positive value representing your gross annual income.}

\item{state_name}{A character. 'OK' for Oklahoma, and 'CA' for California. Default is 'OK'.}
}
\value{
A tibble with the following things: Your federal top marginal rate, state top marginal rate, federal taxes, state taxes, amount paid to social security and medicare, proportion removed from gross income, net income, and what your net income would be if there were no state taxes.
}
\description{
This function takes a numeric value representing your gross income and returns a tibble
with information like net income, the amount you pay in federal taxes and state taxes (possible
states are Oklahoma and California), and other things. Married filing jointly is assumed.
}
\examples{
# What is your net income in Oklahoma if your gross income is $80,000?
taxes(80000)

}
